add validate method to edit contact form
