+++ src/scenes/ServiceMembers/ContactInfo.jsx
@@ -33,7 +33,6 @@ const validateContactForm = (values, form) => {
     const newError = {
       phone_is_preferred: 'Please select a preferred method of contact.',
     };
-    console.log('newe error', newError);
     return newError;
   }
   return errors;
